First working version with blendsplitter
Currently the old view is in a splitter visible on first workspace.

The old code had a lot of direct signals/slots and calls between
DebuggerForm and all the *Viewer code To decouple this the
SignalDispatcher was created as a central way to forward signals to all
views that registers. Also the CPU registers state code now also lives
in the SignalDispatcher.

This means that now we can have multiple DisasmViewers since the code
doesn't rely on one and exactly one disasmView to exist.

As a side effect the QAction that would toggle breakpoints are now
disabled, since that code can no longer as the cursorAdress from the
one-and-only disasmView the old code assumed.

+++ src/DisasmViewer.h
@@ -84,7 +84,7 @@ private:
 	int lineAtPos(const QPoint& pos);
 
 signals:
-	void breakpointToggled(int addr);
+    void breakpointToggled(uint16_t addr);
 };
 
 #endif // DISASMVIEWER_H
